Token auth via cookie (username/email +pwd based)
- added email auth option in case email observed
Not 100% sure whether it's better then having
an explicit EmailOrUsernameAuthBackend atm. Let's give it a try. (at least feeling sort of accomplishment of creating an alternative
only where it's actual, since admin does not need that)
- now both username and email fields should be unique to make authentication
work more or less reliably. Usually thats' a case - both username and email should be unique for a real member